changes to run mipmap gen kernel

+++ executables/cuda_fst/gpu_cull/w_cyls.cpp
@@ -75,7 +75,7 @@ void cleaningScreen(
     );
 
 void downsamplingDepthTexture(
-    float* depthBuffer, 
+    unsigned int* depthBuffer, 
     cudaSurfaceObject_t downsampleSurface, 
     int screenResolutionX, 
     int screenResolutionY,
